+++ compiler-utils/src/testFixtures/java/com/squareup/anvil/compiler/internal/testing/AnvilCompilationMode.kt
@@ -9,6 +9,7 @@ import com.squareup.anvil.compiler.api.CodeGenerator
 public sealed class AnvilCompilationMode(public val analysisBackend: AnalysisBackend) {
   public data class Embedded(
     val codeGenerators: List<CodeGenerator> = emptyList(),
+    val useDagger: Boolean = false,
   ) : AnvilCompilationMode(EMBEDDED)
   public data class Ksp(
     val symbolProcessorProviders: List<SymbolProcessorProvider> = emptyList(),